don't use 'Active' filter; it's deprecated
The active/inactive users filters are deprecated and will be replaced
with a single "Show inactive users" filter; active users will be
returned in all search/filter results. Prefix new user-barcodes with a
'0' to make them findable since it will no longer be possible to
retrieve them with the 'Active' filter. Enter a '0' as a search in order
to show the 'Reset all' button which then allows the form to be
completely cleared prior to subsequent searches.

Refs UIU-400