only choose users that exist
The selector ':nth-of-type(9)' only works if there are at least nine
items to choose from. In testing, there are not so many. 9 is likely a
legacy value from a different environment that had more data but no user
filters and was an effort to avoid selecting the diku_admin user, which
is not a regular user (it has no barcode) among other things, but the
selectors now in place already avoid that account.