don't wait for non-existent rows to appear
Nine wouldn't display with a filter
Nine wouldn't display with a search
And the "active" checkbox is no longer
Though UIU-400 I should not besmirch
Seven, eight, nine!
Help the tester find the line!

It was an itsy bisty teenie weenie list of only fo-ur users
That you got when searched for '0'
It was an itsy bitsy teenie weenie list without a ni-inth user
The test waited. Forever. No joke.