remote exercise test timers
The checkout action in exercise.js would sometimes register as failed
even though it was clear the checkout action itself succeeded (because
subsequent actions relying on an item being checked out were
successful). This meant the test itself, rather than the behavior it was
supposed to be testing, was at fault.

Fixes UITEST-54