wait for elements before clicking them
Oof. `wait(666)`. Really? Timers are notoriously brittle and generate
inconsistent results across different environments. In particular, it
seemed the `xclick()` calls were problematic because there was no
preceeding `wait()` call because the selectors were xpath, not CSS, and
`Nightmare.wait()` expects a CSS selector. Adding the `hasEntries` class
to `<EntrySelector>` should help. Tests with that in place are reliable
so far.

Refs UITEST-55