don't hang tests off `title` attribute
The `title` attribute is being phased out so we need to hang tests off
other attributes.

Additionally, it appears that `xclick` may be dangerous to your health,
or at least needs to be used with caution. In Nightmare, some functions,
like `evaluate()` return promises and therefore _must_ be followed by a
`then(nightmare ... done()).catch()` call that contains the remainder of
the test. Without a `then()`, you can find yourself in a dangling
process that will eventually timeout.

Refs STCOM-296