wait after username entry on checkout
On checkout, it is _sometimes_ the case that if you click the 'Enter'
button very quickly after entering a value in the patron-id field you
get the error message, "User with this id does not exist". I can
sometimes get this in manual testing by typing a username and
immediately hitting the enter key.

Refs UICHKOUT-473