store proxy's barcode, not username
barcodes are likely to be unique across the system, but usernames (or
portions thereof) may turn up in other users' names. Because the
user-search feature includes username, first name, last name, and email,
this means that while the username itself may be unique, a search for it
may turn up multiple rows.

Fixes FOLIO-1657