instantiate new Nightmare instance for tests
Using the passed-in instance of Nightmare leads to weird, weird results
where tests seem to run interleaved with one another and things
generally fall apart in strange, unpredictable ways that ultimately
result in tests that fail when run as part of a suite but succeed when
run individually.

It's like the instance of Nightmare gets tired and just can't keep
itself together. Sadly, this means we can't use Nightmare to collect
coverage data since it's using that single instance that allows us to
generate that data. Alas.

Fixes FOLIO-1660