include app updates
ui-inventory had a bug fix-release, v1.5.2, that should be included in
the platform. It should be included automatically with tilde-versioning,
but we know we want, so let's be specific.

ui-users had a large PR, #660, that incorporated both bug fixes and many
new features that were targeted for Q4, hence a new minor version
release, v2.19.0.