integration tests must accommodate MCL aria changes
Changes to the aria attributes in MCL must be reflected in the
integration tests as well.

Did a little reorganizing too, kebab-casing all tests for consistency
and removing those that now live in their respective repositories
(because they are just regression tests, not integration tests) and do
not need to be in the platform.

Refs STRIPES-596, STCOM-365