correctly restore loan rules
Aw geez yet another "I'm not sure how this ever worked before" commit.
Previously, we saved the default loan rules to a variable defined in
Nightmare from the context of the browser. That ain't right. So now, we
correctly save the rules in the context of nightmare so that we can
restore them when the test finishes.

Coupled with folio-org/ui-circulation/pull/268, this should allow the
loan-renewal test to pass.

Refs STRIPES-602