use correct identifier for plugin's list
folio-org/ui-plugin-find-user/pull/66 changed the way
ui-plugin-find-user provides module information when the plugin is
instantiated. This caused the results-list for a plugin's search to
change from `list-users` to `list-plugin-find-user` because the element
ID is, apparently, constructed from the module ID and we had a bunch of
tests whose selectors needed updating.

Arguably, that PR does The Right Thing by causing the plugin's ID to be
unique instead of matching that of the app it consumes, but it also
highlights the larger issue that plugins really should not be consuming
apps in this manner. It's borderline illegal and just gross.

Refs STRIPES-603