don't click on things in browser scope
avoid clicking links in browser scope. instead, try even harder to
figure out a CSS selector for the thing to be clicked and pass that back
to Nightmare and let Nightmare click it. That seems to make Nightmare
happy, and does a better job of avoiding the problem of Nightmare
getting lost in the never neverland of parallel execution.

Of course, this isn't completely true as the 'renew loan' checkboxes
are, strangely, unclickable in Nightmare's scope. WTF?