use locale-formatted dates
`<DatePicker>` now supports locale-formatted dates. You get these
automatically when using the UI to select a date, but when pasting a
string into the field it must be correctly formatted to match the
locale.

Release v1.6.1 of ui-circulation incorporates the `<DatePicker>` changes
which are reflected in the updated test here that sets a specific locale
and then uses dates formatted for that locale.