better delete/wait handling
Don't wait for edit-button when confirming delete. After the delete is
processed, no loan-item will be selected, hence no edit-button will be
available. The `.wait('#clickable-edit-item')` line was left over from a
previous incarnation of the tests. It didn't cause trouble in other test
environments where the test must breeze past it while still showing the
policy's details, but for some reason it doesn't work that way in CI.

Refs UICIRC-237