clean up exercise test
The exercise test was failing after months of stability. I don't why it
started failing now, but here we are. The work was actually split over a
few PRs, folio-org/stripes-testing/pull/31 and a tiny tweak in
ui-inventory to relabel the new-holdings-button id to something more
sensical.

Here, I replaced `type` with `insert` for adding the user and item
barcodes because the input element was losing focus when the text was
only partially entered. I have no explanation for why this was the case,
but I saw it frequently during failed rounds of tests accompanied by a
warning that "<partial-barcode> could not be found."