use more conservative ~ dep on react (#377)
stripes and the platform both have a hard dep on react but not with the
same restriction; the platform uses `^` and stripes  `~`. Probably we
only want a hard dep in one place. Until that's resolved, we need to at
least use a consistent dependency so we don't get multiple versions of
React in the build when different versions satisfy the different
dependencies (as when 16.9 was released, which is permitted by ^16.8 but
not by ~16.8).

Refs UIIN-678