CIRC-495 overdue policy is required
Circulation rule syntax is not technically part of the API surface, so
this new requirement to have an overdue policy with every rule snuck in
without anybody on the UI side noticing.

These changes provide a dummo policy, `o test-overdue`, in the rules but
do not actually handle creating and removing such a policy in the UI.
We should add that.

Refs CIRC-495