add lost-item policy to circ rules
A backend change now requires including a "lost item" policy reference
in every rule. Who knew? I LOVE it that we don't have a way to version
circ-rule syntax changes in our API. It's totally my favorite thing.
`</sarcasm>`. OK, I get why it doesn't work that way because the API
endpoint just accepts a text-string not a JSON object, but that still
sucks.

All I want for Christmas is an API that accepts circulation rules as
JSON instead of as a string. Or a hippopotamus. A hippopotamus is no
ignormaus and would sing me this sweet serenade:
JSON, JSON, glorious JSON
Nothing quite like for spec'ing the types in
Ugh. POST with me, POST
A string to the host
Until we can send it some glorious JSON.

Alas, I fear somebody is about to punch out my two front teeth and leave
me instead a big old box of rocks.