exercise test must contain valid circ rules
The rules in the circulation policy were missing a lost-item policy.
This meant they were not being correctly saved (which mean the circ
rules test helper needs some work, but that's another story).

Additionally, changes to the handling of close-pane buttons meant we
didn't always see the search results we expected. Adding an explicit
search for the active user to the start of every navigate-to-users
function makes this more reliable.