wait after user search
This test fails pretty consistently after conducting the user search
with an error like
```
Error: .wait() for #list-users[aria-rowcount="2"] timed out after 30000msec
```
which doesn't make any sense; it is literally the same block of code
repeated from other places in the same test. It seems that sometimes,
the results pane doesn't populate although the details pane does, which
is what allows other tests to continue successfully. In the real
non-test-automated world, I have not been able to duplicate this.