FOLIO-2440 create temporary policies
This is a follow-up to #524, applying the same changes to additional
tests.

Use helpers to add/remove overdue fine policy, lost item fee policy, and
patron notice policy in the circulation rules that are created. Checkout
is now smart enough to check all the attributes of circulation policies
and if we just use dummy values, checkout will fail.

Refs FOLIO-2440